finished，用户模块通过测试

+++ server/src/main/java/cn/joker/entity/TaskEntity.java
@@ -215,7 +215,7 @@ public class TaskEntity implements Serializable{
     }
 
     public Double getTotalProgress(){
-        return (double) (completedNumber / expectedNumber);
+        return  ((double)completedNumber / (double)expectedNumber);
     }
 
     public void setTotalProgress(Double totalProgress) {
